feat(user-auth): implement authentication logic
User can sign up with a login and password, and confirm the generated
hash by sending it to discord bot's DM. The bot confirms if there's any
message containing the hashcode and if true, the user is authenticated
and his id is saved in the session.

Every protected by authentication router will verify bot's DM to find
the user that sent the hash code.